requirements.txt
# ---------- Core RAG Framework ----------
llama-index==0.10.68
llama-index-core==0.10.68.post1

# ---------- OpenAI Integration ----------
openai==1.40.3
llama-index-llms-openai==0.1.31
llama-index-embeddings-openai==0.1.11
llama-index-program-openai==0.1.7

# ---------- Data Readers ----------
llama-index-readers-file==0.1.31
pypdf==4.3.1

# ---------- Data Models ----------
pydantic==2.5.3

# ---------- Database & Vector Store ----------
psycopg2-binary==2.9.9
chromadb==0.4.24
sentence-transformers==2.7.0
huggingface-hub==0.19.4

# ---------- Utilities ----------
python-dotenv==1.0.1	# pinned for compatibility
tenacity==8.5.0

# ---------- Numerical (safe for NLP) ----------
numpy==1.26.4

# ---------- SQL connectivity ----------
psycopg2-binary==2.9.9

# ---------- PDF attachments ----------
fpdf==1.7.2
xhtml2pdf==0.2.11
